backend/requirements.txt
fastapi>=0.111
uvicorn[standard]>=0.30
pydantic>=2.6
pydantic-settings>=2.2
sqlalchemy>=2.0
asyncpg>=0.29
psycopg>=3.1
alembic>=1.13
python-jose>=3.3
passlib[bcrypt]>=1.7
python-multipart>=0.0.9

pytest>=8.0
pytest-asyncio>=0.23
httpx>=0.27
